refactor: Raw API 응답과 가공된 DTO 분리
- Client 패키지의 DTO들을 성격에 따라 분리하여 역할을 명확화
  - Raw API 응답: Riot API에서 직접 받는 순수한 응답 데이터
  - 가공된 DTO: 비즈니스 로직을 통해 가공된 데이터 (MatchStats, RankStats, FormattedMatchResponse)
