+++ src/main/java/com/summoner/lolhaeduo/client/riot/RiotClient.java
@@ -1,6 +1,10 @@
 package com.summoner.lolhaeduo.client.riot;
 
-import com.summoner.lolhaeduo.client.dto.*;
+import com.summoner.lolhaeduo.client.dto.api.LeagueEntryResponse;
+import com.summoner.lolhaeduo.client.dto.api.MatchResponse;
+import com.summoner.lolhaeduo.client.dto.api.PuuidResponse;
+import com.summoner.lolhaeduo.client.dto.api.SummonerResponse;
+import com.summoner.lolhaeduo.domain.account.dto.game.FormattedMatchResponse;
 import com.summoner.lolhaeduo.domain.account.enums.AccountRegion;
 import com.summoner.lolhaeduo.domain.account.enums.AccountServer;
 import io.micrometer.core.instrument.MeterRegistry;
