+++ src/main/java/com/summoner/lolhaeduo/domain/account/dto/game/FormattedMatchResponse.java
@@ -1,4 +1,4 @@
-package com.summoner.lolhaeduo.client.dto;
+package com.summoner.lolhaeduo.domain.account.dto.game;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
