+++ src/main/java/com/summoner/lolhaeduo/domain/account/dto/game/MatchStats.java
@@ -1,4 +1,4 @@
-package com.summoner.lolhaeduo.client.dto;
+package com.summoner.lolhaeduo.domain.account.dto.game;
 
 import com.summoner.lolhaeduo.domain.duo.enums.QueueType;
 import lombok.AllArgsConstructor;
