+++ src/main/java/com/summoner/lolhaeduo/domain/account/dto/game/RankStats.java
@@ -1,4 +1,4 @@
-package com.summoner.lolhaeduo.client.dto;
+package com.summoner.lolhaeduo.domain.account.dto.game;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
