+++ src/main/java/com/summoner/lolhaeduo/domain/account/service/RiotGameDataService.java
@@ -1,6 +1,11 @@
 package com.summoner.lolhaeduo.domain.account.service;
 
-import com.summoner.lolhaeduo.client.dto.*;
+import com.summoner.lolhaeduo.client.dto.api.LeagueEntryResponse;
+import com.summoner.lolhaeduo.client.dto.api.PuuidResponse;
+import com.summoner.lolhaeduo.client.dto.api.SummonerResponse;
+import com.summoner.lolhaeduo.domain.account.dto.game.FormattedMatchResponse;
+import com.summoner.lolhaeduo.domain.account.dto.game.MatchStats;
+import com.summoner.lolhaeduo.domain.account.dto.game.RankStats;
 import com.summoner.lolhaeduo.domain.account.entity.Favorite;
 import com.summoner.lolhaeduo.domain.account.repository.FavoriteRepository;
 import com.summoner.lolhaeduo.domain.gamedata.repository.VersionRepository;
